Created Card wrapper component
To avoid code duplication: in styling there are bits which I use several times, so created a wrapper component which I can use instead of the root div to apply that certain styling to other components. In the card component I return a div, which is set up with props.children and props.className so that the styling applied to the children elements will be used too.